% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericModelClasses.R
\docType{class}
\name{CategorModel}
\alias{CategorModel}
\title{R6 class for modeling (fitting and predicting) joint probability for a univariate categorical outcome \code{A[m]}}
\format{An \code{\link{R6Class}} generator object}
\usage{
CategorModel
}
\description{
\code{CategorModel} inherits from \code{\link{GenericModel}} class, defining and modeling a conditional density \eqn{P(A[m]|W,E...)}
 where \code{A[m]} is univariate and categorical. By calling \code{self$new()}, \code{A[m]} will be redefined into number of bins 
 \code{length(levels)} (i.e., number of unique categories in \code{A[m]}). By calling \code{self$fit()}, it fits hazard regressoin
 \code{Bin_A[m][k] ~ W + E} on \code{data} (a \code{\link{DatKeepClass}} class), which is the hazard probaility of the observation 
 of A[m] belongs to bin \code{Bin_A[m][t]}, given covariates \eqn{(W, E)} and that observation doesn't belong to any precedent bins 
 \code{Bin_A[m][1]}, \code{Bin_A[m][2]}, ..., \code{Bin_A[m][k-1]}.
}
\details{
\itemize{
\item{\code{reg}} - .
\item{\code{outvar}} - .
\item{\code{levels}} - Numeric vector of all unique categories in outcome outvar. 
\item{\code{nbins}} - .
\item{\code{bin_nms}} - .
}
}
\section{Methods}{

\describe{
  \item{\code{new(reg, DatKeepClass.g0, ...)}}{Instantiate an new instance of \code{CategorModel} for a univariate categorical outcome \code{A[m]}}
  \item{\code{fit(data)}}{...}
  \item{\code{predict(newdata)}}{...}
  \item{\code{predictAeqa(newdata)}}{...}
}
}

\section{Active Bindings}{

\describe{
  \item{\code{cats}}{...}
}
}

\seealso{
\code{\link{DatKeepClass}}, \code{\link{RegressionClass}}, \code{\link{GenericModel}}, \code{\link{BinaryOutModel}}
}
\keyword{R6}
\keyword{class}
