% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbarDensityModel.R
\name{fitGenericDensity}
\alias{fitGenericDensity}
\title{Define and fit the multivariate conditional density under the user-specified arbitrary intervention function.}
\usage{
fitGenericDensity(data, Anodes, Wnodes, gform = NULL, f_gstar,
  h.gstar_GenericModel = NULL, lbound = 0.01, n_MCsims = 1,
  obs.wts = NULL, rndseed = NULL, verbose = TRUE)
}
\arguments{
\item{data}{\code{data.frame} with named columns, containing \code{Wnodes}, \code{Anode} and \code{Ynode}.}

\item{Anodes}{Column names or indices in \code{data} of outcome variables; exposures can be either binary, categorical or continuous.}

\item{Wnodes}{Column names or indices in \code{data} of baseline covariates. Factors are not currently allowed.}  

\item{gform}{Character vector of regression formula for estimating the conditional density of P(A | W)}

\item{f_gstar}{Either a function or a vector or a matrix/ data frame of counterfactual exposures. See details in function argument 
\code{f_gstar1} in \code{\link{tmleCommunity}}.}

\item{h.gstar_GenericModel}{...}

\item{lbound}{lower bounds on estimated cumulative probabilities for \code{P(A=a|W=w)}, default to 0.01}

\item{n_MCsims}{...}

\item{obs.wts}{...}

\item{rndseed}{...}

\item{verbose}{...}
}
\value{
A named list with 3 items containing the estimation results for:
 \itemize{
 \item \code{h_gstar} - A vector of likelihood prediction for \code{P(A=a|W=w)} where a is generated under the 
  user-specified intervention.
 \item \code{OData.gstar} - A \code{DatKeepClassclass} object,where outcomes are generated under intervention \code{f_gstar}.
 \item \code{genericmodels.gstar} - A \code{GenericModel} class object that defines and models \code{P(A=a|W=w)}.
}
}
\description{
Defines and fits regression models for the conditional density \code{P(A=a|W=w)} where \code{a} is generated under the user-
specified arbitrary (can be static, dynamic or stochastic) intervention function \code{f_gstar}. Note that \code{A} can be 
multivariate \code{(A[1], ..., A[j])} and each of the compoenents A[i] can be either binary, categorical or continuous. 
See detailed description in \code{\link{RegressionClass}}.
}
\examples{
data(indSample.iid.cA.bY_list)
indSample.iid.cA.bY <- indSample.iid.cA.bY_list$indSample.iid.cA.bY
tmleCom_Options(gestimator = "speedglm__glm", maxNperBin = nrow(indSample.iid.cA.bY),
                bin.method = "dhist", nbins = 8)
gvars$verbose <- TRUE  # Print status messages (global setting)

# Define a stochastic intervention
define_f.gstar <- function(data, shift.rate, ...) {
  print("rate of shift: " \%+\% shift.rate)
  shifted.new.A <- data[, "A"] * shift.rate
  return(shifted.new.A)
}
f.gstar <- define_f.gstar(shift.rate = 0.5)

# Under current treatment mechanism g0
h_gN <- fitGenericDensity(data = indSample.iid.cA.bY, Anodes = "A", 
                          Wnodes = c("W1", "W2", "W3", "W4"), 
                          f_gstar = NULL, lbound = 0)$h_gstar
# Under stochastic intervention gstar
h_gstar <- fitGenericDensity(data = indSample.iid.cA.bY, Anodes = "A",
                             Wnodes = c("W1", "W2", "W3", "W4"), 
                             f_gstar = f.gstar, lbound = 0)$h_gstar
}
\seealso{
\code{\link{tmleCom_Options}}, \code{\link{DatKeepClass}}, \code{\link{RegressionClass}}, \code{\link{GenericModel}}, 
 \code{\link{ContinModel}}, \code{\link{CategorModel}}, \code{\link{tmleCommunity}}
}
