% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenericModelClasses.R
\docType{class}
\name{GenericModel}
\alias{GenericModel}
\title{Generic R6 class for modeling (fitting and predicting) P(A=a | W=w, E=e) where A can be multivariate (A[1], ..., A[j])}
\format{An \code{\link{R6Class}} generator object}
\usage{
GenericModel
}
\description{
\code{GenericModel} defines and models the conditional density \code{P(A=a|W=w,E=e)}, where \code{a} are generated under \code{g_star}
 or \code{g_0}. By calling \code{self$new(reg)}, it utilizes estimation options defined in \code{\link{RegressionClass}} class, and 
 automatically factorizes \code{P(A|W,E)} into an entire tree of binary regression models, where a new instance of \code{\link{BinaryOutModel}}
 class will be initialized for each binary regression. By calling \code{self$fit(data)} and \code{self$predict(newdata)}, where \code{data} and    
 \code{newdata} are \code{\link{DatKeepClass}} class objects, it fits \code{P(A|W,E)} and predicts \code{P(A=1|W=w,E=e)}, where values of 
 {\code{w}, \code{e}} are from \code{newdata}. Moreover, it predicts likelihood function P(A=a| W=w,E=e) through \code{self$predictAeqa(newdata)},
 where {\code{a}, \code{w}, \code{e}} are from \code{newdata} (also a \code{\link{DatKeepClass}} class).
}
\details{
\itemize{
\item{\code{n_regs}} - Total number of regression models.
\item{\code{parfit_allowed}} - Logical. If TRUE, allow parallel computing to fit multivariate outvar.
}
}
\section{Methods}{

\describe{
  \item{\code{new(reg, ...)}}{Use \code{reg} (a \code{\link{RegressionClass}} class object) to instantiate an new object of \code{GenericModel}}
  \item{\code{length}}{...}
  \item{\code{getPsAsW.models}}{Get all model objects (one model object per outcome var A[j])}
  \item{\code{getcumprodAeqa}}{Get joint prob as a vector of the cumulative prod over j for P(A[j]=a[j]|W,E)}
  \item{\code{fit(data, savespace = TRUE)}}{savespace is Logical. If TRUE, wipe out all internal data when doing many stacked regressions}
  \item{\code{copy.fit(Generic.Model)}}{...}
  \item{\code{predict(newdata, savespace = TRUE)}}{...}
  \item{\code{predictAeqa(newdata, savespace = TRUE, wipeProb = TRUE, ...)}}{wipeProb is logical argument for \code{self$wipe.alldat}. If 
    FALSE, vectors of probA1 & probAeqa will be kept.}
}
}

\section{Active Bindings}{

\describe{
  \item{\code{wipe.alldat(wipeProb = TRUE)}}{...}
}
}

\keyword{R6}
\keyword{class}
