% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmleCommunity-package.R
\docType{data}
\name{indSample.iid.cA.cY_list}
\alias{indSample.iid.cA.cY_list}
\title{An Example of a Non-Hierarchical Data Containing a Continuous Exposure with a Continuous Outcome.}
\format{A data frame with 10,000 independent observations (rows) and 6 variables:
\describe{
  \item{W1}{binary baseline covariate with \eqn{P(W1=1) = 0.5}}
  \item{W2}{binary baseline covariate with \eqn{P(W2=1) = 0.3}}
  \item{W3}{continuous normal baseline covariate with \eqn{\mu} = 0 and \eqn{\sigma} = 0.25}
  \item{W4}{continuous uniform baseline covariate with \code{min=0} and \code{max=1}}
  \item{A}{continuous normal exposure where its mean depends on individual's baseline covariate values in \code{(W1, W2, W3, W4)}}
  \item{Y}{continuous normal outcome where its mean depends on individual's baseline covariate and exposure values in (\code{W1}, 
    \code{W2}, \code{W3}, \code{W4}, \code{A})}
}}
\source{
\url{https://github.com/chizhangucb/tmleCommunity/blob/master/tests/dataGeneration/get.iid.dat.Acont.R}
}
\usage{
data(indSample.iid.cA.cY_list)
}
\description{
Simulated (non-hierarchical) dataset containing 10,000 i.i.d. observations, with each row \code{i} consisting of measured baseline 
 covariates (\code{W1}, \code{W2}, \code{W3} and \code{W4}), continuous exposure (\code{A}) and continous outcome (\code{Y}). 
 The baseline covariates \code{W1}, \code{W2}, \code{W3} and \code{W4} were sampled as i.i.d., while the value of exposure \code{A} 
 for each observation \code{i} was drawn conditionally on the value of \code{i}'s four baseline covariates. Besides, the continuous
 outcome \code{Y} for each observation depends on \code{i}'s baseline covariates and exposure values in (\code{W1[i]},\code{W2[i]},
 \code{W3[i]}, \code{W4[i]}, \code{A[i]}). The following section provides more details regarding individual variables in simulated 
 data.
}
\examples{
data(indSample.iid.cA.cY_list)
indSample.iid.cA.cY <- indSample.iid.cA.cY_list$indSample.iid.cA.cY
# True mean of outcome under intervention g0
psi0.Y <- indSample.iid.cA.cY_list$psi0.Y  
# True mean of outcoem under stochastic intervention gstar
psi0.Ygstar <- indSample.iid.cA.cY_list$psi0.Ygstar  
# truncated bound used in sampling A* under gstar (in data generating mechanism)
indSample.iid.cA.cY_list$truncBD  
# shift value used in sampling A* under gstar 
indSample.iid.cA.cY_list$shift.val
}
\keyword{datasets}
