% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmleCommunity-package.R
\docType{data}
\name{indSample.ind.bA.bY.rareJ1_list}
\alias{indSample.ind.bA.bY.rareJ1_list}
\title{An Example of a Non-Hierarchical Data Containing a Binary Exposure with a Rare Binary Outcome (Independent Case-Control)}
\format{A data frame with 2,000 independent observations (rows), containing 1000 cases and 1000 controls, and 6 variables, :
\describe{
  \item{W1} {continuous uniform baseline covariate with \code{min=0} and \code{max=1}}
  \item{W2} {continuous normal baseline covariate with \eqn{\mu} = 0 and \eqn{\sigma} = 0.3}
  \item{W3} {binary baseline covariate with \eqn{P(W2=1) = 0.5}}
  \item{W4} {binary baseline covariate with \eqn{P(W2=1) = 0.5}}
  \item{A} {binary exposure that depends on baseline covariate values in \code{(W1, W2, W3, W4)}}
  \item{Y} {binary outcome that depends on baseline covariate and exposure values in (\code{W1, W2, W3, W4, A})}
    
}}
\usage{
data(indSample.ind.bA.bY.rareJ1_list)
}
\description{
Simulated (non-hierarchical) dataset containing 2,000 i.i.d. observations, with each row \code{i} consisting of 4 measured baseline 
 covariates (\code{W1}, \code{W2}, \code{W3} and \code{W4}), 1 binary exposure (\code{A}) and 1 binary outcome (\code{Y}) that
 defines case or control status. The baseline covariates \code{W1}, \code{W2}, \code{W3} and \code{W4} were sampled as i.i.d., 
 while the exposure \code{A} for each observation \code{i} depends on \code{i}'s four baseline covariates. Similarly, the outcome
 \code{Y} for each observation depends on \code{i}'s baseline covariates and exposure values. Moreover, we can also describe the 
 case-control design as first sampling \eqn{1} case \eqn{(W_1^1, W_2^1, W_3^1, W_4^1, A^1)} from the conditional distribution of 
 \eqn{(W_1, W_2, W_3, W_4, A)}, given Y = 1. One then samples \eqn{J} controls \eqn{(W_1^{0,j}, W_2^{0,j}, W_3^{0,j}, 
 W_4^{0,j}, A^{0,j})} from \eqn{(W_1, W_2, W_3, W_4, A)}, given Y = 0, \eqn{j=1,...,J}. Thus, the cluster containing one case 
 and \code{J} controls is considered the experimental unit. Finally one gets \eqn{nC} cases and \eqn{nCo} controls with 
 \eqn{J=nC/nCo}, where \eqn{J} can be used effectively in observation weights. The following section provides more details
 regarding individual variables in simulated data.
}
\examples{
data(indSample.ind.bA.bY.rareJ1_list)
indSample.ind.bA.bY.rareJ1 <- indSample.ind.bA.bY.rareJ1_list$indSample.ind.bA.bY.rareJ1
head(indSample.ind.bA.bY.rareJ1_list$obs.wt.J1)  # Assigned weights to each observations
indSample.ind.bA.bY.rareJ1_list$q0  # 0.013579 True prevalence probability
indSample.ind.bA.bY.rareJ1_list$psi0.Y  # 0.012662 True ATE
indSample.ind.bA.bY.rareJ1_list$J  # 1 The ratio of number of controls to cases
}
\keyword{datasets}
