% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmleCommunity-package.R
\docType{data}
\name{indSample.ind.bA.bY.rareJ2_list}
\alias{indSample.ind.bA.bY.rareJ2_list}
\title{An Example of a Non-Hierarchical Data Containing a Binary Exposure with a Rare Outcome (Independent Case-Control J = 2)}
\format{A data frame with 3,000 independent observations (rows), containing 1000 cases and 2000 controls, and 6 variables:}
\usage{
data(indSample.ind.bA.bY.rareJ2_list)
}
\description{
Simulated (non-hierarchical) dataset containing 3,000 i.i.d. observations. The data structure of \code{indSample.ind.bA.bY.rareJ2}
is identical to this of \code{indSample.ind.bA.bY.rareJ1}, except that now the ratio of the number of controls to the number of 
case \eqn{J} is 2.
}
\keyword{datasets}
