% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloSimClass.R
\docType{class}
\name{MonteCarloSimClass}
\alias{MonteCarloSimClass}
\title{R6 class for evaluating different plug-in estimators via for Monte-Carlo resampling where new exposures are generated 
under the user-specified arbitrary intervention function g^{*}.}
\format{An \code{\link{R6Class}} generator object}
\usage{
MonteCarloSimClass
}
\description{
\code{MonteCarloSimClass} only resamples \code{A} under the intervention function \code{g_star}, never \code{W} or \code{E}.
 For each MC simulation, it firstly treats \code{model.Q.init} as the fitted models for \code{E[Y|A,W,E]}, then estimate 
 \code{psi_n} using Monte-Carlo integration. i.e., average of \code{n} predicted \code{E(Y|A=a, W=w,E=e)} where \code{a} is 
 a vector of \code{n} new exposures randomly drawn under \code{g_star}. Take as many iterations as needed until convergence occurs.
}
\details{
\itemize{
\item{\code{OData.ObsP0}} - .
\item{\code{OData.gstar}} - .
\item{\code{model.Q.init}} - .
\item{\code{model.Q.star.cov}} - .
\item{\code{model.Q.star.int}} - .
\item{\code{QY.init}} - .
\item{\code{QY.star.cov}} - .
\item{\code{QY.star.int}} - .
\item{\code{nobs}} - Number of observations in the observed data frame.
\item{\code{p}} - Number of Monte-Carlo simulations performed.
}
}
\section{Methods}{

\describe{
  \item{\code{new(OData.ObsP0, OData.gstar, ...)}}{...}
  \item{\code{get.gcomp(m.Q.init)}}{...}
  \item{\code{get.tmleCov(model.Q.star.cov, model.h.fit)}}{...}
  \item{\code{get.tmleInt(model.Q.star.int)}}{...}
  \item{\code{get.fiW()}}{...}
}
}

\keyword{R6}
\keyword{class}
