% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print_tmleCom_opts}
\alias{print_tmleCom_opts}
\title{Print Current Option Settings for \code{tmleCommunity}}
\usage{
print_tmleCom_opts()
}
\value{
Invisibly returns a list of \code{tmleCommunity} options.
}
\description{
Print Current Option Settings for \code{tmleCommunity}
}
\seealso{
\code{\link{tmleCom_Options}}
}
