% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{tmleCom_Options}
\alias{tmleCom_Options}
\title{Setting all possible \code{tmleCommunity} options}
\usage{
tmleCom_Options(Qestimator = c("speedglm__glm", "glm__glm", "h2o__ensemble",
  "SuperLearner"), gestimator = c("speedglm__glm", "glm__glm",
  "h2o__ensemble", "SuperLearner"), bin.method = c("equal.mass", "equal.len",
  "dhist"), nbins = 10, maxncats = 10, maxNperBin = 500, parfit = FALSE,
  poolContinVar = FALSE, savetime.fit.hbars = TRUE,
  h2ometalearner = "h2o.glm.wrapper", h2olearner = "h2o.glm.wrapper",
  CVfolds = 5, g.SL.library = c("SL.glm", "SL.step", "SL.glm.interaction"))
}
\arguments{
\item{Qestimator}{A string specifying default estimator for outcome mechanism model fitting. 
The default estimator is \code{"speedglm__glm"}, which estimates regressions with \code{\link[speedglm]{speedglm.wfit}}; 
Estimator \code{"glm__glm"} uses \code{\link[stats]{glm.fit}};
Estimator \code{"h2o__ensemble"} implements the super learner ensemble (stacking) algorithm using the H2O R interface; 
Estimator \code{"SuperLearner"} implements the super learner prediction methods.
Note that if \code{"h2o__ensemble"} fails, it falls back on {"SuperLearner"}. If \code{"SuperLearner"} fails, 
it falls back on {"speedglm__glm"}. If \code{"speedglm__glm"} fails, it falls back on {"glm__glm"}.}

\item{gestimator}{A string specifying default estimator for exposure mechanism fitting. It has the same options as \code{Qestimator}.}

\item{bin.method}{Specify the method for choosing bins when discretizing the conditional continuous exposure variable \code{A}.
The default method is \code{"equal.len"}, which partitions the range of \code{A} into equal length \code{nbins} intervals.
Method \code{"equal.mass"} provides a data-adaptive selection of the bins based on equal mass/ area, i.e., each bin will contain 
approximately the same number of observations as otheres. Method \code{"dhist"} uses a combination of the above two approaches,
see Denby and Mallows "Variations on the Histogram" (2009) for more details. Note that argument \code{maxNperBin} will control
the maximum number of observations in each bin.}

\item{nbins}{Set the default number of bins when discretizing a continous variable under setting \code{bin.method = "equal.len"}. 
When method is \code{"equal.mass"}, \code{nbins} will be set as the maximum of the default \code{nbins} and the nearest integer
of \code{nobs}/\code{maxNperBin}, where \code{nobs} is the total number of observations in the input data. 
If setting to \code{NA}, \code{nbins} is defined as the nearest integer of \code{nobs}/\code{maxNperBin}.}

\item{maxncats}{Set the maximum number of unique categories a categorical variable \code{A[j]} can have. If \code{A[j]} has more
unique categories, it is automatically considered a continuous variable. Default to 10.}

\item{maxNperBin}{Set the maximum number of observations in each bin when discretizing a continuous variable \code{A[j]} (applies 
directly when \code{bin.method = "equal.mass"} and indirectly when \code{bin.method = "equal.len"}, but \code{nbins = NA}).}

\item{parfit}{Logical. If \code{TRUE}, perform parallel regression fits and predictions for discretized continuous variables by 
functions \code{foreach} and \code{dopar} in \code{foreach} package. Note that it requires registering a parallel backend prior 
to running \code{tmleCommunity} function, e.g., using \code{doParallel} R package and running 
\code{registerDoParallel(cores = ncores)} for \code{ncores} parallel jobs. Default to \code{FALSE}.}

\item{poolContinVar}{Logical. If \code{TRUE}, when fitting a model for binirized continuous variable, pool bin indicators across
all bins and fit one pooled regression. Default to \code{FALSE}.}

\item{savetime.fit.hbars}{Logical. If \code{TRUE}, it will skip exposure mechanism P(A|W,E) under g0 & gstar fits and predictions
when \code{f.gstar = NULL} and \code{TMLE.targetStep = "tmle.intercept"}, and directly set \code{h_gstar_h_gN = 1} for each 
observation. Default to \code{TRUE}.}

\item{h2ometalearner}{A string to pass to \code{\link{h2o.ensemble}}, specifying the prediction algorithm used to learn the optimal 
combination of the base learners. Supports both h2o and SuperLearner wrapper functions. Default to "h2o.glm.wrapper".}

\item{h2olearner}{A string or character vector to pass to \code{\link{h2o.ensemble}}, naming the prediction algorithm(s) used to train
the base models for the ensemble. The functions must have the same format as the h2o wrapper functions. Default to "h2o.glm.wrapper".}

\item{CVfolds}{Set the number of splits for the V-fold cross-validation step to pass to \code{\link{SuperLearner}} and 
\code{\link{h2o.ensemble}}. Default to 5.}

\item{g.SL.library}{A string or character vector of prediction algorithms to pass to \code{\link{SuperLearner}}. Default to 
c("SL.glm", "SL.step", "SL.glm.interaction"). See \code{SuperLearner::listWrappers()} for more available algorithms.}
}
\value{
Invisibly returns a list with old option settings.
}
\description{
Additional options that control \code{tmleCommunity} package
}
